runner: send context to Destination writers
This commit introduces a `DestinationContext` object that captures
additional information about the running job, for Destinations that need
it. For now, it includes the Job ID, and the Source and Destination
options provided.

This enables Destinations to vary their behavior based on the context
they're running in. For example, our Balsa Glider plugin will use this
to read the organization and source IDs, ensuring it writes synced data
to the proper org and source.

BREAKING CHANGE: Job ID moved from the first argument to a `jobId`
property on the context object.